+++ src/metrics/Metric.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 abstract public class Metric
 {
-  abstract public int compare(List<Double> reference, List<Double> vector1, List<Double> vector2);
+  abstract public int compare(List<Float> reference, List<Float> vector1, List<Float> vector2);
 }
