+++ src/metrics/TaxicabMetric.java
@@ -4,10 +4,10 @@ import java.util.List;
 
 public class TaxicabMetric extends Metric
 {
-  public int compare(List<Double> reference, List<Double> vector1, List<Double> vector2)
+  public int compare(List<Float> reference, List<Float> vector1, List<Float> vector2)
   {
-    Double sum1 = 0.0;
-    Double sum2 = 0.0;
+    Float sum1 = 0.0f;
+    Float sum2 = 0.0f;
     for(Integer i = 0; i < reference.size(); ++i)
     {
       sum1+= Math.abs(reference.get(i)-vector1.get(i));
